Traverse FiledDecls in checking reachability from the entry point
When a input file is like:

```go
import (
    gobytes "bytes"
)

type T struct {
    ident *gobytes.Buffer
}
```

and the `gobytes` is not used in the rest of the file,  the `gobytes` in
the FieldDecl will remain in the combined source.  The renaming is failed
because the `gobytes` is not marked as reachable from the entry point.

This commit will fix it.

Close: #5